chore: create a generate data function
